EOEPCA-939 #comment update ADES chart files
Keep the ADES chart sql and services in-sync with the upstream ZOO-Project

+++ charts/ades/files/zoo-services/services/UndeployProcess.py
@@ -1,3 +1,28 @@
+#
+# Author : Blasco Brauzzi, Fabrice Brito, Frank Löschau
+#
+# Copyright 2023 Terradue. All rights reserved.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the
+# "Software"), to deal in the Software without restriction, including with
+# out limitation the rights to use, copy, modify, merge, publish,
+# distribute, sublicense, and/or sell copies of the Software, and to
+# permit persons to whom the Software is furnished to do so, subject to
+# the following conditions:
+#
+# The above copyright notice and this permission notice shall be included
+# in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+# OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
+# CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+# TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+#
+
 import os
 import shutil
 import json
@@ -62,24 +87,30 @@ class UndeployService(object):
         return process_identifier
 
 
-
-
     def remove_service(self):
-        import psycopg2
-        import psycopg2.extensions
-        psycopg2.extensions.register_type(psycopg2.extensions.UNICODE)
-        conn = psycopg2.connect("host=%s port=%s dbname=%s user=%s password=%s" % (self.conf["metadb"]["host"], self.conf["metadb"]["port"], self.conf["metadb"]["dbname"], self.conf["metadb"]["user"], self.conf["metadb"]["password"]))
-        cur = conn.cursor()
-        cur.execute("DELETE FROM collectiondb.ows_process WHERE identifier='%s'" % (self.get_process_identifier()))
-        conn.commit()
-        conn.close()
+
+        if "metadb" in self.conf:
+            import psycopg2
+            import psycopg2.extensions
+            psycopg2.extensions.register_type(psycopg2.extensions.UNICODE)
+            conn = psycopg2.connect("host=%s port=%s dbname=%s user=%s password=%s" % (self.conf["metadb"]["host"], self.conf["metadb"]["port"], self.conf["metadb"]["dbname"], self.conf["metadb"]["user"], self.conf["metadb"]["password"]))
+            cur = conn.cursor()
+            if "auth_env" in self.conf:
+                self.user=self.conf["auth_env"]["user"]
+            else:
+                self.user="anonymous"
+            cur.execute("DELETE FROM collectiondb.ows_process WHERE identifier='%s' AND user_id=(select id from public.users where name=$q$%s$q$)" % (self.get_process_identifier(),self.user))
+            conn.commit()
+            conn.close()
+
         service_folder = os.path.join(self.zooservices_folder, self.service_identifier)
         if os.path.isdir(service_folder):
             shutil.rmtree(service_folder)
 
-        service_configuration_file = f"{service_folder}.zcfg"
-        if os.path.exists(service_configuration_file):
-            os.remove(service_configuration_file)
+        for i in [".zcfg",".json"]:
+            service_configuration_file = f"{service_folder}"+i
+            if os.path.exists(service_configuration_file):
+                os.remove(service_configuration_file)
